feat: reward group and sample
