+++ Assets/Sample/Scripts/RewardSystem.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 67cf04f5f508438abbba89d3a895e59b
+timeCreated: 1754155851
